feat(openscenario_interpreter): mv ScenarioError, add catch InitActions

+++ openscenario/openscenario_interpreter/src/syntax/trigger.cpp
@@ -51,7 +51,7 @@ auto Trigger::evaluate() -> Object
 
 auto Trigger::activeConditionGroupIndex() const -> int
 {
-  int index{0};
+  size_t index{0};
   for (auto it = begin(); it != end(); ++it, ++index) {
     ConditionGroup const & cgroup = *it;
     bool all_conditions_true = std::all_of(
